android: move core views
