+++ android/src/main/java/com/s77rt/rtndatepicker/RTNDatePickerView.kt
@@ -1,41 +1,22 @@
 package com.s77rt.rtndatepicker
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.requiredWidth
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.material3.BasicAlertDialog
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.DatePicker
 import androidx.compose.material3.DatePickerDefaults
 import androidx.compose.material3.DatePickerDialog
-import androidx.compose.material3.DatePickerFormatter
-import androidx.compose.material3.DisplayMode
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.LocalMinimumInteractiveComponentSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.material3.TimePicker
-import androidx.compose.material3.TimePickerColors
 import androidx.compose.material3.TimePickerDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.DialogProperties
 import java.time.Instant
 import java.time.ZoneId
 
@@ -261,125 +242,6 @@ fun RTNDatePickerView(
                     headline = datePickerHeadline,
                     showModeToggle = showModeToggle,
                 )
-            }
-        }
-    }
-}
-
-// Based on compose/material3/material3/src/commonMain/kotlin/androidx/compose/material3/DatePicker.kt
-@OptIn(ExperimentalMaterial3Api::class)
-@Suppress("FunctionName")
-@Composable
-fun DatePickerTitle(
-    title: String,
-    displayMode: DisplayMode,
-    contentColor: Color,
-) {
-    val modifier = Modifier.padding(PaddingValues(start = 24.dp, end = 12.dp, top = 16.dp))
-
-    if (title.isEmpty()) {
-        DatePickerDefaults.DatePickerTitle(
-            displayMode = displayMode,
-            modifier = modifier,
-            // TODO: Pass contentColor after upgrading material 3 to v1.4.0+
-            // contentColor = contentColor,
-        )
-    } else {
-        Text(
-            text = title,
-            modifier = modifier,
-            color = contentColor,
-        )
-    }
-}
-
-// Based on compose/material3/material3/src/commonMain/kotlin/androidx/compose/material3/DatePicker.kt
-@OptIn(ExperimentalMaterial3Api::class)
-@Suppress("FunctionName")
-@Composable
-fun DatePickerHeadline(
-    headline: String,
-    selectedDateMillis: Long?,
-    displayMode: DisplayMode,
-    dateFormatter: DatePickerFormatter,
-    contentColor: Color,
-) {
-    val modifier = Modifier.padding(PaddingValues(start = 24.dp, end = 12.dp, bottom = 12.dp))
-
-    if (headline.isEmpty()) {
-        DatePickerDefaults.DatePickerHeadline(
-            selectedDateMillis = selectedDateMillis,
-            displayMode = displayMode,
-            dateFormatter = dateFormatter,
-            modifier = modifier,
-            // TODO: Pass contentColor after upgrading material 3 to v1.4.0+
-            // contentColor = contentColor,
-        )
-    } else {
-        Text(
-            text = headline,
-            modifier = modifier,
-            color = contentColor,
-            maxLines = 1,
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Suppress("FunctionName")
-@Composable
-fun TimePickerTitle(
-    title: String,
-    contentColor: Color,
-) {
-    Text(
-        text = if (title.isEmpty()) "Select time" else title,
-        modifier = Modifier.padding(bottom = 20.dp),
-        style = MaterialTheme.typography.labelMedium,
-        color = contentColor,
-    )
-}
-
-// Based on compose/material3/material3/src/androidMain/kotlin/androidx/compose/material3/DatePickerDialog.android.kt
-@OptIn(ExperimentalMaterial3Api::class)
-@Suppress("FunctionName")
-@Composable
-fun TimePickerDialog(
-    onDismissRequest: () -> Unit,
-    confirmButton: @Composable () -> Unit,
-    dismissButton: @Composable () -> Unit,
-    colors: TimePickerColors,
-    title: (@Composable () -> Unit)?,
-    content: @Composable ColumnScope.() -> Unit,
-) {
-    BasicAlertDialog(
-        onDismissRequest = onDismissRequest,
-        modifier = Modifier.wrapContentHeight(),
-        properties = DialogProperties(usePlatformDefaultWidth = false),
-    ) {
-        Surface(
-            shape = DatePickerDefaults.shape,
-            tonalElevation = DatePickerDefaults.TonalElevation,
-            color = colors.containerColor,
-        ) {
-            Column(modifier = Modifier.padding(24.dp), verticalArrangement = Arrangement.SpaceBetween) {
-                if (title != null) {
-                    title()
-                }
-                Box(Modifier.weight(1f, fill = false)) {
-                    this@Column.content()
-                }
-                Box(
-                    modifier =
-                        Modifier.align(Alignment.End),
-                ) {
-                    Row {
-                        CompositionLocalProvider(LocalMinimumInteractiveComponentSize provides 0.dp) {
-                            dismissButton()
-                            confirmButton()
-                        }
-                    }
-                }
             }
         }
     }
